RichF1 - Formula 1 Website with Database
This application is a 3rd year project for Advanced Internet
Technologies. It is a website with a relational database underneath,
that stores the drivers, teams, sponsors and race tracks in Formula 1,
allowing the user to add, edit and delete entries in the database.

The application also has user profiles and allows registration with
0Auth for Facebook and Google profiles.
The application also includes 2-step verification by text message.